refactor formatting code into its own module
