+++ pixeltable/dataframe.py
@@ -1,22 +1,16 @@
 from __future__ import annotations
 
-import base64
 import copy
 import hashlib
-import io
 import json
 import logging
-import mimetypes
 import traceback
 from pathlib import Path
 from typing import List, Optional, Any, Dict, Generator, Tuple, Set
 
-import PIL.Image
-import cv2
 import pandas as pd
 import pandas.io.formats.style
 import sqlalchemy as sql
-from PIL import Image
 
 import pixeltable.catalog as catalog
 import pixeltable.exceptions as excs
@@ -26,20 +20,11 @@ from pixeltable.catalog import is_valid_identifier
 from pixeltable.env import Env
 from pixeltable.plan import Planner
 from pixeltable.type_system import ColumnType
-from pixeltable.utils.http_server import get_file_uri
-import numpy as np
+from pixeltable.utils.formatter import PixeltableFormatter
+
 
 __all__ = ['DataFrame']
-
 _logger = logging.getLogger('pixeltable')
-
-
-def _create_source_tag(file_path: str) -> str:
-    src_url = get_file_uri(Env.get().http_address, file_path)
-    mime = mimetypes.guess_type(src_url)[0]
-    # if mime is None, the attribute string would not be valid html.
-    mime_attr = f'type="{mime}"' if mime is not None else ''
-    return f'<source src="{src_url}" {mime_attr} />'
 
 
 class DataFrameResultSet:
@@ -47,13 +32,14 @@ class DataFrameResultSet:
         self._rows = rows
         self._col_names = col_names
         self._col_types = col_types
+        self._formatter = PixeltableFormatter(len(self._rows), len(self._col_names), Env.get().http_address)
         self._formatters = {
-            ts.ImageType: self._format_img,
-            ts.VideoType: self._format_video,
-            ts.AudioType: self._format_audio,
-            ts.DocumentType: self._format_document,
-            ts.ArrayType: self._format_array,
-            ts.StringType: self._format_string,
+            ts.ImageType: self._formatter._format_img,
+            ts.VideoType: self._formatter._format_video,
+            ts.AudioType: self._formatter._format_audio,
+            ts.DocumentType: self._formatter._format_document,
+            ts.ArrayType: self._formatter._format_array,
+            ts.StringType: self._formatter._format_string,
 #            ts.JsonType: self._format_json,
         }
 
@@ -75,7 +61,7 @@ class DataFrameResultSet:
             for col_name, col_type in zip(self._col_names, self._col_types)
             if col_type.__class__ in self._formatters
         }
-        return self.to_pandas().to_html(formatters=formatters, escape=True, index=False)
+        return self.to_pandas().to_html(formatters=formatters, escape=False, index=False)
 
     def __str__(self) -> str:
         return self.to_pandas().to_string()
@@ -89,119 +75,6 @@ class DataFrameResultSet:
 
     def _row_to_dict(self, row_idx: int) -> Dict[str, Any]:
         return {self._col_names[i]: self._rows[row_idx][i] for i in range(len(self._col_names))}
-
-    # Formatters
-    def _format_img(self, img: Image.Image) -> str:
-        """
-        Create <img> tag for Image object.
-        """
-        assert isinstance(img, Image.Image), f'Wrong type: {type(img)}'
-        # Try to make it look decent in a variety of display scenarios
-        if len(self._rows) > 1:
-            width = 240  # Multiple rows: display small images
-        elif len(self._col_names) > 1:
-            width = 480  # Multiple columns: display medium images
-        else:
-            width = 640  # A single image: larger display
-        with io.BytesIO() as buffer:
-            img.save(buffer, 'jpeg')
-            img_base64 = base64.b64encode(buffer.getvalue()).decode()
-            return f"""
-            <div class="pxt_image" style="width:{width}px;">
-                <img src="data:image/jpeg;base64,{img_base64}" width="{width}" />
-            </div>
-            """
-
-    def _format_video(self, file_path: str) -> str:
-        thumb_tag = ''
-        # Attempt to extract the first frame of the video to use as a thumbnail,
-        # so that the notebook can be exported as HTML and viewed in contexts where
-        # the video itself is not accessible.
-        # TODO(aaron-siegel): If the video is backed by a concrete external URL,
-        # should we link to that instead?
-        video_reader = cv2.VideoCapture(str(file_path))
-        if video_reader.isOpened():
-            status, img_array = video_reader.read()
-            if status:
-                img_array = cv2.cvtColor(img_array, cv2.COLOR_BGR2RGB)
-                thumb = PIL.Image.fromarray(img_array)
-                with io.BytesIO() as buffer:
-                    thumb.save(buffer, 'jpeg')
-                    thumb_base64 = base64.b64encode(buffer.getvalue()).decode()
-                    thumb_tag = f'poster="data:image/jpeg;base64,{thumb_base64}"'
-            video_reader.release()
-        if len(self._rows) > 1:
-            width = 320
-        elif len(self._col_names) > 1:
-            width = 480
-        else:
-            width = 800
-        return f"""
-        <div class="pxt_video" style="width:{width}px;">
-            <video controls width="{width}" {thumb_tag}>
-                {_create_source_tag(file_path)}
-            </video>
-        </div>
-        """
-
-    def _format_array(self, arr: List[Any]) -> str:
-        arr = np.array(arr)
-        return np.array2string(arr, precision=3, threshold=16, separator=',', edgeitems=6)
-
-    def _format_string(self, string: str) -> str:
-        if len(string) > 250:
-            return f'{string[:120]} ...... {string[-120:]}'
-        return string
-
-    def _format_json(self, obj : Any) -> str:
-        if isinstance(obj, list):
-            return self._format_array(obj)
-        elif isinstance(obj, dict):
-            for key, value in obj.items():
-                fmt_value = _format_json(value)
-
-                obj[key] = self._format_json(value)
-            return json.dumps(obj, indent=4)
-
-    def _format_document(self, file_path: str) -> str:
-        max_width = max_height = 320
-        # by default, file path will be shown as a link
-        inner_element = file_path
-        # try generating a thumbnail for different types and use that if successful
-        if file_path.lower().endswith('.pdf'):
-            try:
-                import fitz
-
-                doc = fitz.open(file_path)
-                p = doc.get_page_pixmap(0)
-                while p.width > max_width or p.height > max_height:
-                    # shrink(1) will halve each dimension
-                    p.shrink(1)
-                data = p.tobytes(output='jpeg')
-                thumb_base64 = base64.b64encode(data).decode()
-                img_src = f'data:image/jpeg;base64,{thumb_base64}'
-                inner_element = f"""
-                    <img style="object-fit: contain; border: 1px solid black;" src="{img_src}" />
-                """
-            except:
-                logging.warning(f'Failed to produce PDF thumbnail {file_path}. Make sure you have PyMuPDF installed.')
-
-        return f"""
-        <div class="pxt_document" style="width:{max_width}px;">
-            <a href="{get_file_uri(Env.get().http_address, file_path)}">
-                {inner_element}
-            </a>
-        </div>
-        """
-
-    def _format_audio(self, file_path: str) -> str:
-        return f"""
-        <div class="pxt_audio">
-            <audio controls>
-                {_create_source_tag(file_path)}
-            </audio>
-        </div>
-        """
 
     def __getitem__(self, index: Any) -> Any:
         if isinstance(index, str):
