+++ pixeltable/utils/formatter.py
@@ -1,0 +1,131 @@
+import base64
+import numpy as np
+import PIL
+import PIL.Image as Image
+import io
+import cv2
+import mimetypes
+from typing import Any
+
+from pixeltable.utils.http_server import get_file_uri
+import logging
+
+_logger = logging.getLogger('pixeltable')
+
+def _create_source_tag(http_address: str, file_path: str) -> str:
+    src_url = get_file_uri(http_address, file_path)
+    mime = mimetypes.guess_type(src_url)[0]
+    # if mime is None, the attribute string would not be valid html.
+    mime_attr = f'type="{mime}"' if mime is not None else ''
+    return f'<source src="{src_url}" {mime_attr} />'
+
+class PixeltableFormatter:
+    def __init__(self, num_rows: int, num_cols: int, http_address: str):
+        self.num_rows = num_rows
+        self.num_cols = num_cols
+        self.http_address = http_address
+
+    def _format_img(self, img: Image.Image) -> str:
+        """
+        Create <img> tag for Image object.
+        """
+        assert isinstance(img, Image.Image), f'Wrong type: {type(img)}'
+        # Try to make it look decent in a variety of display scenarios
+        if self.num_rows > 1:
+            width = 240  # Multiple rows: display small images
+        elif self.num_cols > 1:
+            width = 480  # Multiple columns: display medium images
+        else:
+            width = 640  # A single image: larger display
+        with io.BytesIO() as buffer:
+            img.save(buffer, 'jpeg')
+            img_base64 = base64.b64encode(buffer.getvalue()).decode()
+            return f"""
+            <div class="pxt_image" style="width:{width}px;">
+                <img src="data:image/jpeg;base64,{img_base64}" width="{width}" />
+            </div>
+            """
+
+    def _format_video(self, file_path: str) -> str:
+        thumb_tag = ''
+        # Attempt to extract the first frame of the video to use as a thumbnail,
+        # so that the notebook can be exported as HTML and viewed in contexts where
+        # the video itself is not accessible.
+        # TODO(aaron-siegel): If the video is backed by a concrete external URL,
+        # should we link to that instead?
+        video_reader = cv2.VideoCapture(str(file_path))
+        if video_reader.isOpened():
+            status, img_array = video_reader.read()
+            if status:
+                img_array = cv2.cvtColor(img_array, cv2.COLOR_BGR2RGB)
+                thumb = PIL.Image.fromarray(img_array)
+                with io.BytesIO() as buffer:
+                    thumb.save(buffer, 'jpeg')
+                    thumb_base64 = base64.b64encode(buffer.getvalue()).decode()
+                    thumb_tag = f'poster="data:image/jpeg;base64,{thumb_base64}"'
+            video_reader.release()
+        if self.num_rows > 1:
+            width = 320
+        elif self.num_cols > 1:
+            width = 480
+        else:
+            width = 800
+        return f"""
+        <div class="pxt_video" style="width:{width}px;">
+            <video controls width="{width}" {thumb_tag}>
+                {_create_source_tag(self.http_address, file_path)}
+            </video>
+        </div>
+        """
+
+    def _format_array(self, arr: list[Any]) -> str:
+        arr = np.array(arr)
+        return np.array2string(arr, precision=3, threshold=16, separator=',', edgeitems=6)
+
+    def _format_string(self, string: str) -> str:
+        if len(string) > 250:
+            return f'{string[:120]} ...... {string[-120:]}'
+        return string
+
+    def _format_json(self, obj: Any) -> str:
+        pass
+
+    def _format_document(self, file_path: str) -> str:
+        max_width = max_height = 320
+        # by default, file path will be shown as a link
+        inner_element = file_path
+        # try generating a thumbnail for different types and use that if successful
+        if file_path.lower().endswith('.pdf'):
+            try:
+                import fitz
+
+                doc = fitz.open(file_path)
+                p = doc.get_page_pixmap(0)
+                while p.width > max_width or p.height > max_height:
+                    # shrink(1) will halve each dimension
+                    p.shrink(1)
+                data = p.tobytes(output='jpeg')
+                thumb_base64 = base64.b64encode(data).decode()
+                img_src = f'data:image/jpeg;base64,{thumb_base64}'
+                inner_element = f"""
+                    <img style="object-fit: contain; border: 1px solid black;" src="{img_src}" />
+                """
+            except:
+                logging.warning(f'Failed to produce PDF thumbnail {file_path}. Make sure you have PyMuPDF installed.')
+
+        return f"""
+        <div class="pxt_document" style="width:{max_width}px;">
+            <a href="{get_file_uri(self.http_address, file_path)}">
+                {inner_element}
+            </a>
+        </div>
+        """
+
+    def _format_audio(self, file_path: str) -> str:
+        return f"""
+        <div class="pxt_audio">
+            <audio controls>
+                {_create_source_tag(self.http_address, file_path)}
+            </audio>
+        </div>
+        """
